refactored node name, exported python shell definition, fixed python.js failing on process starts empty checking if null
