+++ quantum/nodes/grovers-algo/grovers.html
@@ -1,5 +1,5 @@
 <script type="text/javascript">
-  RED.nodes.registerType("grovers-algo", {
+  RED.nodes.registerType("grovers", {
     category: "quantum",
     color: "#737177",
     defaults: {
@@ -18,7 +18,7 @@
   });
 </script>
 
-<script type="text/html" data-template-name="grovers-algo">
+<script type="text/html" data-template-name="grovers">
   <div class="form-row">
     <label for="node-input-name"><i class="fa fa-tag"></i> Name</label>
     <input type="text" id="node-input-name" placeholder="Name" />
